refactor(ui/mobile/liquidGlass): extract LiquidGlassBlur styled-components to separate file
- Add SimpleLoadingContainer, SimpleNativeBlurContainer, SimpleCSSBlurContainer to LiquidGlassBlur.js
- Update LiquidGlassBlur.jsx to import styled components from separate style file
- Remove inline styled component definitions from component file
- Maintain existing liquid glass functionality while improving code organization
- Follow project pattern of separating styles from component logic
