+++ app/imports/ui/mobile/liquidGlass/components/LiquidGlassBlur.jsx
@@ -1,11 +1,15 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import { useNativeBlur } from '../../hooks/useNativeBlur';
+import {
+  SimpleLoadingContainer,
+  SimpleNativeBlurContainer,
+  SimpleCSSBlurContainer,
+} from '../styles/LiquidGlassBlur';
 
 /**
  * LiquidGlassBlur Component
- * 
+ *
  * Provides iOS 26 Liquid Glass blur effects with automatic fallback to CSS
  * Uses native UIVisualEffectView when available for optimal performance
  */
@@ -41,7 +45,7 @@ const LiquidGlassBlur = ({
   // Determine if we should use native blur
   useEffect(() => {
     mountedRef.current = true;
-    
+
     if (!isLoading) {
       setUseNative(nativeSupported && window.cordova && window.Meteor?.isCordova);
     }
@@ -81,10 +85,10 @@ const LiquidGlassBlur = ({
       };
 
       const id = await createBlurView(blurOptions);
-      
+
       if (mountedRef.current) {
         setBlurId(id);
-        
+
         if (onBlurReady) {
           onBlurReady({ blurId: id, useNative: true });
         }
@@ -109,7 +113,7 @@ const LiquidGlassBlur = ({
   // Handle visibility changes
   const handleVisibilityChange = useCallback((visible) => {
     setIsVisible(visible);
-    
+
     if (useNative && blurId) {
       setBlurVisibility(blurId, visible).catch(console.error);
     }
@@ -136,16 +140,16 @@ const LiquidGlassBlur = ({
 
   if (isLoading) {
     return (
-      <LoadingContainer ref={containerRef} className={className} style={style} {...props}>
+      <SimpleLoadingContainer ref={containerRef} className={className} style={style} {...props}>
         {children}
-      </LoadingContainer>
+      </SimpleLoadingContainer>
     );
   }
 
   if (useNative) {
     // Native blur - transparent container that content renders over the native blur
     return (
-      <NativeBlurContainer
+      <SimpleNativeBlurContainer
         ref={containerRef}
         className={className}
         style={style}
@@ -154,13 +158,13 @@ const LiquidGlassBlur = ({
         {...props}
       >
         {children}
-      </NativeBlurContainer>
+      </SimpleNativeBlurContainer>
     );
   }
 
   // CSS fallback
   return (
-    <CSSBlurContainer
+    <SimpleCSSBlurContainer
       ref={containerRef}
       className={className}
       style={style}
@@ -173,7 +177,7 @@ const LiquidGlassBlur = ({
       {...props}
     >
       {children}
-    </CSSBlurContainer>
+    </SimpleCSSBlurContainer>
   );
 };
 
@@ -181,7 +185,7 @@ LiquidGlassBlur.propTypes = {
   children: PropTypes.node,
   blurStyle: PropTypes.oneOf([
     'systemMaterial',
-    'systemThinMaterial', 
+    'systemThinMaterial',
     'systemThickMaterial',
     'systemChromeMaterial',
     'systemUltraThinMaterial',
@@ -204,65 +208,5 @@ LiquidGlassBlur.propTypes = {
   fallbackStyle: PropTypes.oneOf(['light', 'dark', 'tinted']),
   onBlurReady: PropTypes.func
 };
-
-// Styled components
-const LoadingContainer = styled.div`
-  position: relative;
-  background: rgba(255, 255, 255, 0.1);
-  opacity: 0.6;
-`;
-
-const NativeBlurContainer = styled.div`
-  position: ${props => props.position};
-  background: transparent;
-  opacity: ${props => props.isVisible ? 1 : 0};
-  transition: opacity 0.3s ease;
-  
-  /* Ensure content is visible over native blur */
-  z-index: 10;
-`;
-
-const CSSBlurContainer = styled.div`
-  position: ${props => props.position};
-  opacity: ${props => props.isVisible ? 1 : 0};
-  transition: ${props => props.animated ? 'all 0.3s ease' : 'none'};
-  
-  /* CSS backdrop-filter fallback */
-  background: ${props => {
-    const alpha = props.intensity * 0.7;
-    switch (props.blurStyle) {
-      case 'dark':
-        return `rgba(0, 0, 0, ${alpha})`;
-      case 'tinted':
-        return `rgba(120, 120, 128, ${alpha})`;
-      default:
-        return `rgba(255, 255, 255, ${alpha})`;
-    }
-  }};
-  
-  backdrop-filter: blur(${props => props.intensity * 20}px);
-  -webkit-backdrop-filter: blur(${props => props.intensity * 20}px);
-  
-  ${props => props.floating && `
-    border-radius: 16px;
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.2);
-    margin: 8px;
-  `}
-  
-  /* Fallback for browsers without backdrop-filter support */
-  @supports not (backdrop-filter: blur(1px)) {
-    background: ${props => {
-      const alpha = Math.min(props.intensity * 0.9, 0.95);
-      switch (props.blurStyle) {
-        case 'dark':
-          return `rgba(0, 0, 0, ${alpha})`;
-        case 'tinted':
-          return `rgba(120, 120, 128, ${alpha})`;
-        default:
-          return `rgba(255, 255, 255, ${alpha})`;
-      }
-    }};
-  }
-`;
 
 export default LiquidGlassBlur;
