+++ app/imports/ui/mobile/liquidGlass/styles/LiquidGlassBlur.js
@@ -37,22 +37,22 @@ const glassShimmer = keyframes`
 export const BlurContainer = styled.div`
   position: ${props => props.position || 'relative'};
   overflow: hidden;
-  
+
   ${props => props.animated && css`
     transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
   `}
-  
+
   ${props => props.floating && css`
     border-radius: 20px;
     margin: 16px;
-    box-shadow: 
+    box-shadow:
       0 8px 32px rgba(0, 0, 0, 0.12),
       0 2px 8px rgba(0, 0, 0, 0.08),
       inset 0 1px 0 rgba(255, 255, 255, 0.1);
-    
+
     /* iOS 26 Liquid Glass border */
     border: 0.5px solid rgba(255, 255, 255, 0.2);
-    
+
     /* Subtle floating animation */
     animation: ${floatingAnimation} 4s ease-in-out infinite;
   `}
@@ -62,7 +62,7 @@ export const BlurContainer = styled.div`
 export const NativeBlurContainer = styled(BlurContainer)`
   background: transparent;
   z-index: 10;
-  
+
   /* Ensure content is properly positioned over native blur */
   > * {
     position: relative;
@@ -73,15 +73,15 @@ export const NativeBlurContainer = styled(BlurContainer)`
 // CSS fallback blur styles
 export const CSSBlurContainer = styled(BlurContainer)`
   --blur-intensity: ${props => (props.intensity || 1) * 20}px;
-  
+
   background: ${props => getBlurBackground(props.blurStyle, props.intensity)};
   backdrop-filter: blur(var(--blur-intensity));
   -webkit-backdrop-filter: blur(var(--blur-intensity));
-  
+
   ${props => props.animated && css`
     animation: ${fadeInBlur} 0.4s ease-out;
   `}
-  
+
   /* iOS 26 style glass effect */
   ${props => props.iosStyle && css`
     background: linear-gradient(
@@ -89,7 +89,7 @@ export const CSSBlurContainer = styled(BlurContainer)`
       ${getBlurBackground(props.blurStyle, props.intensity * 0.8)} 0%,
       ${getBlurBackground(props.blurStyle, props.intensity * 1.2)} 100%
     );
-    
+
     /* Glass shimmer overlay */
     &::before {
       content: '';
@@ -109,11 +109,11 @@ export const CSSBlurContainer = styled(BlurContainer)`
       pointer-events: none;
     }
   `}
-  
+
   /* Fallback for browsers without backdrop-filter */
   @supports not (backdrop-filter: blur(1px)) {
     background: ${props => getBlurBackground(props.blurStyle, Math.min(props.intensity * 0.95, 0.98))};
-    
+
     &::before {
       content: '';
       position: absolute;
@@ -132,7 +132,7 @@ export const CSSBlurContainer = styled(BlurContainer)`
 export const LoadingContainer = styled(BlurContainer)`
   background: rgba(255, 255, 255, 0.1);
   opacity: 0.6;
-  
+
   &::after {
     content: '';
     position: absolute;
@@ -146,7 +146,7 @@ export const LoadingContainer = styled(BlurContainer)`
     border-radius: 50%;
     animation: spin 1s linear infinite;
   }
-  
+
   @keyframes spin {
     0% { transform: translate(-50%, -50%) rotate(0deg); }
     100% { transform: translate(-50%, -50%) rotate(360deg); }
@@ -161,7 +161,7 @@ export const ToolbarBlurContainer = styled(BlurContainer)`
   right: 0;
   height: ${props => props.height || '80px'};
   padding-bottom: env(safe-area-inset-bottom);
-  
+
   ${props => props.floating && css`
     bottom: 16px;
     left: 16px;
@@ -169,7 +169,7 @@ export const ToolbarBlurContainer = styled(BlurContainer)`
     border-radius: 24px;
     height: ${props.height || '60px'};
   `}
-  
+
   /* Ensure toolbar is above other content */
   z-index: 1000;
 `;
@@ -179,7 +179,7 @@ export const BlurContent = styled.div`
   position: relative;
   z-index: 2;
   padding: ${props => props.padding || '16px'};
-  
+
   /* Ensure text is readable over blur */
   color: ${props => {
     switch (props.blurStyle) {
@@ -192,7 +192,7 @@ export const BlurContent = styled.div`
         return 'var(--text-primary, rgba(0, 0, 0, 0.9))';
     }
   }};
-  
+
   /* Text shadow for better readability */
   text-shadow: ${props => {
     switch (props.blurStyle) {
@@ -207,7 +207,7 @@ export const BlurContent = styled.div`
 // Helper function to get blur background color
 function getBlurBackground(blurStyle, intensity) {
   const alpha = Math.min(intensity * 0.7, 0.9);
-  
+
   switch (blurStyle) {
     case 'dark':
       return `rgba(0, 0, 0, ${alpha})`;
@@ -226,7 +226,7 @@ function getBlurBackground(blurStyle, intensity) {
 // Export utility functions
 export const BlurUtils = {
   getBlurBackground,
-  
+
   // Get appropriate text color for blur style
   getTextColor: (blurStyle) => {
     switch (blurStyle) {
@@ -239,26 +239,86 @@ export const BlurUtils = {
         return 'rgba(0, 0, 0, 0.9)';
     }
   },
-  
+
   // Check if backdrop-filter is supported
   supportsBackdropFilter: () => {
-    return CSS.supports('backdrop-filter', 'blur(1px)') || 
+    return CSS.supports('backdrop-filter', 'blur(1px)') ||
            CSS.supports('-webkit-backdrop-filter', 'blur(1px)');
   },
-  
+
   // Get optimal blur intensity for device
   getOptimalIntensity: (baseIntensity = 1.0) => {
     // Reduce intensity on older/slower devices
     const performanceMemory = navigator.deviceMemory || 4;
     const hardwareConcurrency = navigator.hardwareConcurrency || 4;
-    
+
     if (performanceMemory < 4 || hardwareConcurrency < 4) {
       return baseIntensity * 0.7;
     }
-    
+
     return baseIntensity;
   }
 };
+
+// Additional containers for component variants
+export const SimpleLoadingContainer = styled.div`
+  position: relative;
+  background: rgba(255, 255, 255, 0.1);
+  opacity: 0.6;
+`;
+
+export const SimpleNativeBlurContainer = styled.div`
+  position: ${props => props.position};
+  background: transparent;
+  opacity: ${props => props.isVisible ? 1 : 0};
+  transition: opacity 0.3s ease;
+
+  /* Ensure content is visible over native blur */
+  z-index: 10;
+`;
+
+export const SimpleCSSBlurContainer = styled.div`
+  position: ${props => props.position};
+  opacity: ${props => props.isVisible ? 1 : 0};
+  transition: ${props => props.animated ? 'all 0.3s ease' : 'none'};
+
+  /* CSS backdrop-filter fallback */
+  background: ${props => {
+    const alpha = props.intensity * 0.7;
+    switch (props.blurStyle) {
+      case 'dark':
+        return `rgba(0, 0, 0, ${alpha})`;
+      case 'tinted':
+        return `rgba(120, 120, 128, ${alpha})`;
+      default:
+        return `rgba(255, 255, 255, ${alpha})`;
+    }
+  }};
+
+  backdrop-filter: blur(${props => props.intensity * 20}px);
+  -webkit-backdrop-filter: blur(${props => props.intensity * 20}px);
+
+  ${props => props.floating && `
+    border-radius: 16px;
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.2);
+    margin: 8px;
+  `}
+
+  /* Fallback for browsers without backdrop-filter support */
+  @supports not (backdrop-filter: blur(1px)) {
+    background: ${props => {
+      const alpha = Math.min(props.intensity * 0.9, 0.95);
+      switch (props.blurStyle) {
+        case 'dark':
+          return `rgba(0, 0, 0, ${alpha})`;
+        case 'tinted':
+          return `rgba(120, 120, 128, ${alpha})`;
+        default:
+          return `rgba(255, 255, 255, ${alpha})`;
+      }
+    }};
+  }
+`;
 
 export default {
   BlurContainer,
@@ -267,5 +327,8 @@ export default {
   LoadingContainer,
   ToolbarBlurContainer,
   BlurContent,
-  BlurUtils
+  BlurUtils,
+  SimpleLoadingContainer,
+  SimpleNativeBlurContainer,
+  SimpleCSSBlurContainer
 };
